+++ tests/resources/__init__.py
@@ -1,3 +1,7 @@
+import pytest
+
+from hy._compat import PYODIDE
+
 in_init = "chippy"
 
 
@@ -7,6 +11,12 @@ def kwtest(*args, **kwargs):
 
 def function_with_a_dash():
     pass
+
+
+can_test_async = not PYODIDE
+async_test = pytest.mark.skipif(
+    not can_test_async, reason="`asyncio.run` not implemented"
+)
 
 
 class AsyncWithTest: