+++ tests/test_bin.py
@@ -12,7 +12,12 @@ from pathlib import Path
 
 import pytest
 
-from hy._compat import PY3_9, PYPY
+from hy._compat import PY3_9, PYODIDE, PYPY
+
+if PYODIDE:
+    pytest.skip(
+        '`subprocess.Popen` not implemented on Pyodide',
+        allow_module_level = True)
 
 
 def pyr(s=""):